+++ packages/expo-file-system/src/FileSystem.ts
@@ -12,6 +12,13 @@ export class Paths extends PathUtilities {
   }
 
   /**
+   * A property containing the bundle directory – the directory where assets bundled with the application are stored.
+   */
+  static get bundle() {
+    return new Directory(ExpoFileSystem.bundleDirectory);
+  }
+
+  /**
    * A property containing the document directory – a place to store files that are safe from being deleted by the system.
    */
   static get document() {
@@ -99,7 +106,7 @@ export class File extends ExpoFileSystem.FileSystemFile implements Blob {
     return bytes.buffer as ArrayBuffer;
   }
 
-  stream(): ReadableStream<Uint8Array> {
+  stream(): ReadableStream<Uint8Array<ArrayBuffer>> {
     return this.readableStream();
   }