+++ packages/expo-ui/ios/HStackView.swift
@@ -30,11 +30,13 @@ internal final class HStackViewProps: ExpoSwiftUI.ViewProps, CommonViewModifierP
   @Field var fixedSize: Bool?
   @Field var frame: FrameOptions?
   @Field var padding: PaddingOptions?
+  @Field var testID: String?
+  @Field var modifiers: ModifierArray?
+
   @Field var spacing: Double?
   @Field var useTapGesture: Bool?
   @Field var alignment: VerticalAlignmentOptions?
   @Field var backgroundColor: Color?
-  @Field var testID: String?
   var onTap = EventDispatcher()
 }