+++ packages/expo-ui/CHANGELOG.md
@@ -10,6 +10,20 @@
 
 ### 💡 Others
 
+## 0.2.0-alpha.2 — 2025-08-16
+
+_This version does not introduce any user-facing changes._
+
+## 0.2.0-alpha.1 — 2025-08-15
+
+### 🛠 Breaking changes
+
+- Merged `swift-ui-primitives` to `swift-ui` and now every component should explicitly wrap with a `<Host>`. ([#38866](https://github.com/expo/expo/pull/38866) by [@kudo](https://github.com/kudo))
+
+### 🎉 New features
+
+- [Android] Add compose modifier support. ([#38630](https://github.com/expo/expo/pull/38630) by [@aleqsio](https://github.com/aleqsio))
+
 ## 0.2.0-alpha.0 — 2025-08-13
 
 ### 🎉 New features
@@ -27,6 +41,7 @@
 - [iOS] Add `testID` property to views. ([#37919](https://github.com/expo/expo/pull/37919) by [@mateoguzmana](https://github.com/mateoguzmana))
 - Add `SecureField` component. ([#37642](https://github.com/expo/expo/pull/37642) by [@aleqsio](https://github.com/aleqsio))
 - [Android] Add `AlertDialog` component. ([#38266](https://github.com/expo/expo/pull/38266) by [@mateoguzmana](https://github.com/mateoguzmana))
+- Added SwiftUI view modifiers support. ([#38543](https://github.com/expo/expo/pull/38543) by [@kudo](https://github.com/kudo))
 
 ### 🐛 Bug fixes